+++ src/packages/media/media/index.ts
@@ -2,7 +2,7 @@ import './components/index.js';
 
 export * from './repository/index.js';
 export * from './workspace/index.js';
-export * from './tracked-reference/index.js';
+export * from './reference/index.js';
 export * from './components/index.js';
 export * from './entity.js';
 export * from './utils/index.js';
