+++ src/packages/media/media/reference/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './media-reference.repository.js';
