+++ src/packages/media/media/reference/repository/media-reference.server.data.ts
@@ -1,0 +1,31 @@
+import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
+import { MediaResource } from '@umbraco-cms/backoffice/external/backend-api';
+import { tryExecuteAndNotify } from '@umbraco-cms/backoffice/resources';
+
+/**
+ * @export
+ * @class UmbMediaReferenceServerDataSource
+ * @implements {RepositoryDetailDataSource}
+ */
+export class UmbMediaReferenceServerDataSource {
+	#host: UmbControllerHost;
+
+	/**
+	 * Creates an instance of UmbMediaReferenceServerDataSource.
+	 * @param {UmbControllerHost} host
+	 * @memberof UmbMediaReferenceServerDataSource
+	 */
+	constructor(host: UmbControllerHost) {
+		this.#host = host;
+	}
+
+	/**
+	 * Fetches the item for the given id from the server
+	 * @param {Array<string>} ids
+	 * @return {*}
+	 * @memberof UmbMediaReferenceServerDataSource
+	 */
+	async getReferenceById(id: string, skip = 0, take = 20) {
+		return await tryExecuteAndNotify(this.#host, MediaResource.getMediaByIdReferencedBy({ id, skip, take }));
+	}
+}
