--- src/packages/media/media/tracked-reference/repository/index.ts
@@ -1,1 +1,0 @@
-export * from './media-tracked-reference.repository.js';
