--- src/packages/media/media/tracked-reference/repository/media-tracked-reference.repository.ts
@@ -1,29 +1,0 @@
-import { UmbMediaTrackedReferenceServerDataSource } from './media-tracked-reference.server.data.js';
-import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
-import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
-
-export class UmbMediaTrackedReferenceRepository extends UmbControllerBase {
-	#trackedReferenceSource: UmbMediaTrackedReferenceServerDataSource;
-
-	constructor(host: UmbControllerHost) {
-		super(host);
-		this.#trackedReferenceSource = new UmbMediaTrackedReferenceServerDataSource(this);
-	}
-
-	async requestTrackedReference(unique: string, skip = 0, take = 20) {
-		if (!unique) throw new Error(`unique is required`);
-		return this.#trackedReferenceSource.getTrackedReferenceById(unique, skip, take);
-	}
-
-	async requestTrackedReferenceDescendantsFromParentUnique(parentUnique: string, skip = 0, take = 20) {
-		if (!parentUnique) throw new Error(`unique is required`);
-		return this.#trackedReferenceSource.getTrackedReferenceDescendantsByParentId(parentUnique, skip, take);
-	}
-
-	async requestTrackedReferenceItems(uniques: string[], skip = 0, take = 20) {
-		if (!uniques) throw new Error(`unique is required`);
-		return this.#trackedReferenceSource.getTrackedReferenceItem(uniques, skip, take);
-	}
-}
-
-export default UmbMediaTrackedReferenceRepository;
